Refactor current implementations of IScopeManager to inherit from ScopeManagerBase so all the implementation is identical except for how the Active scope is stored and retrieved.
