+++ src/Datadog.Trace.AspNet/AspNetScopeManager.cs
@@ -4,24 +4,12 @@ using Datadog.Trace.Logging;
 
 namespace Datadog.Trace.AspNet
 {
-    internal class AspNetScopeManager : IScopeManager
+    internal class AspNetScopeManager : ScopeManagerBase
     {
-        private static readonly Vendors.Serilog.ILogger Log = DatadogLogging.For<AspNetScopeManager>();
-
         private readonly string _name = "__Datadog_Scope_Current__" + Guid.NewGuid();
         private readonly AsyncLocalCompat<Scope> _activeScopeFallback = new AsyncLocalCompat<Scope>();
 
-        public event EventHandler<SpanEventArgs> SpanOpened;
-
-        public event EventHandler<SpanEventArgs> SpanActivated;
-
-        public event EventHandler<SpanEventArgs> SpanDeactivated;
-
-        public event EventHandler<SpanEventArgs> SpanClosed;
-
-        public event EventHandler<SpanEventArgs> TraceEnded;
-
-        public Scope Active
+        public override Scope Active
         {
             get
             {
@@ -33,66 +21,17 @@ namespace Datadog.Trace.AspNet
 
                 return _activeScopeFallback.Get();
             }
-        }
 
-        public Scope Activate(Span span, bool finishOnClose)
-        {
-            var newParent = Active;
-            var scope = new Scope(newParent, span, this, finishOnClose);
-            var scopeOpenedArgs = new SpanEventArgs(span);
-
-            SpanOpened?.Invoke(this, scopeOpenedArgs);
-
-            SetScope(scope);
-
-            if (newParent != null)
+            set
             {
-                SpanDeactivated?.Invoke(this, new SpanEventArgs(newParent.Span));
+                var httpContext = HttpContext.Current;
+                if (httpContext != null)
+                {
+                    httpContext.Items[_name] = value;
+                }
+
+                _activeScopeFallback.Set(value);
             }
-
-            SpanActivated?.Invoke(this, scopeOpenedArgs);
-
-            return scope;
-        }
-
-        public void Close(Scope scope)
-        {
-            var current = Active;
-            var isRootSpan = scope.Parent == null;
-
-            if (current == null || current != scope)
-            {
-                // This is not the current scope for this context, bail out
-                return;
-            }
-
-            // if the scope that was just closed was the active scope,
-            // set its parent as the new active scope
-            SetScope(current.Parent);
-            SpanDeactivated?.Invoke(this, new SpanEventArgs(current.Span));
-
-            if (!isRootSpan)
-            {
-                SpanActivated?.Invoke(this, new SpanEventArgs(current.Parent.Span));
-            }
-
-            SpanClosed?.Invoke(this, new SpanEventArgs(scope.Span));
-
-            if (isRootSpan)
-            {
-                TraceEnded?.Invoke(this, new SpanEventArgs(scope.Span));
-            }
-        }
-
-        private void SetScope(Scope scope)
-        {
-            var httpContext = HttpContext.Current;
-            if (httpContext != null)
-            {
-                httpContext.Items[_name] = scope;
-            }
-
-            _activeScopeFallback.Set(scope);
         }
     }
 }
