+++ src/Datadog.Trace/AsyncLocalScopeManager.cs
@@ -3,71 +3,20 @@ using Datadog.Trace.Logging;
 
 namespace Datadog.Trace
 {
-    internal class AsyncLocalScopeManager : IScopeManager
+    internal class AsyncLocalScopeManager : ScopeManagerBase
     {
-        private static readonly Vendors.Serilog.ILogger Log = DatadogLogging.GetLogger(typeof(AsyncLocalScopeManager));
-
         private readonly AsyncLocalCompat<Scope> _activeScope = new AsyncLocalCompat<Scope>();
 
-        public event EventHandler<SpanEventArgs> SpanOpened;
-
-        public event EventHandler<SpanEventArgs> SpanActivated;
-
-        public event EventHandler<SpanEventArgs> SpanDeactivated;
-
-        public event EventHandler<SpanEventArgs> SpanClosed;
-
-        public event EventHandler<SpanEventArgs> TraceEnded;
-
-        public Scope Active => _activeScope.Get();
-
-        public Scope Activate(Span span, bool finishOnClose)
+        public override Scope Active
         {
-            var newParent = Active;
-            var scope = new Scope(newParent, span, this, finishOnClose);
-            var scopeOpenedArgs = new SpanEventArgs(span);
-
-            SpanOpened?.Invoke(this, scopeOpenedArgs);
-
-            _activeScope.Set(scope);
-
-            if (newParent != null)
+            get
             {
-                SpanDeactivated?.Invoke(this, new SpanEventArgs(newParent.Span));
+                return _activeScope.Get();
             }
 
-            SpanActivated?.Invoke(this, scopeOpenedArgs);
-
-            return scope;
-        }
-
-        public void Close(Scope scope)
-        {
-            var current = Active;
-            var isRootSpan = scope.Parent == null;
-
-            if (current == null || current != scope)
+            set
             {
-                // This is not the current scope for this context, bail out
-                SpanClosed?.Invoke(this, new SpanEventArgs(scope.Span));
-                return;
-            }
-
-            // if the scope that was just closed was the active scope,
-            // set its parent as the new active scope
-            _activeScope.Set(scope.Parent);
-            SpanDeactivated?.Invoke(this, new SpanEventArgs(scope.Span));
-
-            if (!isRootSpan)
-            {
-                SpanActivated?.Invoke(this, new SpanEventArgs(scope.Parent.Span));
-            }
-
-            SpanClosed?.Invoke(this, new SpanEventArgs(scope.Span));
-
-            if (isRootSpan)
-            {
-                TraceEnded?.Invoke(this, new SpanEventArgs(scope.Span));
+                _activeScope.Set(value);
             }
         }
     }
